==> RNA_seq/DataAnalysis/07.scRNA/02.Cluster.Rmd <==
---
title: "Cluster: Clustering the cells by their relative abundance of Genes"
date: "Created: 2021-02-01; modified: `r Sys.Date()`"
output: 
  html_notebook:
    codes: hide
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(warning = F, message = F)
library(dplyr)
library(Seurat)
library(ggplot2)

# rm(list = ls())
options(stringsAsFactors = F)
options(future.globals.maxSize = 4000 * 1024^2)

colors <- c("#A6CEE3", "#1F78B4", "#08306B", "#B2DF8A", "#006D2C", "#8E0152",
         "#DE77AE", "#CAB2D6", "#6A3D9A", "#FB9A99", "#E31A1C", "#B15928",
         "#619CFF","#FF67A4","#00BCD8", "#EE2B2B", "#2D6BB4")
```


### load data 
```{r}
data.integrated <- readRDS("../../Result/sc_profile/hsp_normal/data.normal.RDS")
```

```{r}
table(data.integrated$Individual)
```


### precluster
```{r}
data.integrated <- RunPCA(object = data.integrated, verbose = FALSE)
data.integrated <- RunTSNE(data.integrated, reduction = "pca", dims = 1:30)
data.integrated <- RunUMAP(data.integrated, reduction = "pca", dims = 1:30)

cowplot::plot_grid( 
        DimPlot(data.integrated,
        reduction = "umap",
        label = TRUE,
        label.size = 6), 
        DimPlot(data.integrated,
        reduction = "umap",
        label = TRUE,
        label.size = 6, 
        split.by = "Batch")+NoLegend(),
        ncol = 2)
```


### Clustering cells based on top PCs (metagenes)
```{r}
DimHeatmap(data.integrated, 
           reduction = "pca",
           dims = 1:9, 
           cells = 500, 
           balanced = TRUE)
```

* Printing out the most variable genes driving PCs
```{r}
print(x = data.integrated[["pca"]], 
      dims = 1:10, 
      nfeatures = 5)
```

* Plot the elbow plot
```{r}
ElbowPlot(object = data.integrated, 
          ndims = 40)+
	geom_vline(xintercept = 30, linetype=2)+ 
	geom_hline(yintercept = 2.5, linetype=2)
```


### Cluster the cells
```{r}
# Determine the K-nearest neighbor graph
# data.integrated <- FindNeighbors(object = data.integrated,
#                                  dims = 1:30)

data.integrated <- FindNeighbors(object = data.integrated,
                                 k.param = 30,
                                 dims = 1:10)
                                
# Determine the clusters for various resolutions                                
data.integrated <- FindClusters(object = data.integrated,
                               resolution = c(0.4, 0.6, 0.8, 1.0, 1.4))
```


### choose the middle of the resolution: umap

* integrated_snn_res.0.4
```{r}
Idents(object = data.integrated) <- "integrated_snn_res.0.4"
DimPlot(data.integrated,
        reduction = "umap",  
        label = TRUE,
        label.size = 6)
```


* integrated_snn_res.0.6
```{r}
Idents(object = data.integrated) <- "integrated_snn_res.0.6"
DimPlot(data.integrated,
        reduction = "umap",
        label = TRUE,
        label.size = 8)
```

* integrated_snn_res.0.8
```{r}
Idents(object = data.integrated) <- "integrated_snn_res.0.8"
DimPlot(data.integrated,
        reduction = "umap",
        label = TRUE,
        label.size = 6)
```

* integrated_snn_res.1
```{r}
Idents(object = data.integrated) <- "integrated_snn_res.1"
DimPlot(data.integrated,
        reduction = "umap",
        label = TRUE,
        label.size = 6)
```

* integrated_snn_res.1.4
```{r}
Idents(object = data.integrated) <- "integrated_snn_res.1.4"
DimPlot(data.integrated,
        reduction = "umap",
        label = TRUE,
        label.size = 6)
```

### output 

* 0.4 resolution to check the quality control metrics and known markers
```{r}
if(dir.exists("../../Result/sc_profile/hsp_normal/")){
  dir.create("../../Result/sc_profile/hsp_normal/")
}

Idents(object = data.integrated) <- "integrated_snn_res.1"
saveRDS(data.integrated, "../../Result/sc_profile/hsp_normal/data.normal.cluster.RDS", compress = TRUE)
```


### version
```{r}
sessionInfo()
```
